requirements.txt
asgiref==3.3.1
DateTime==4.3
Django==3.1.4
django-crispy-forms==1.11.1
numpy==1.20.1
pandas==1.2.2
protobuf==3.15.2
psycopg2==2.8.6
python-dateutil==2.8.1
six==1.15.0
sqlparse==0.4.1
zope.interface==5.2.0
